refactor layout to include header and stack
