+++ src/components/index.js
@@ -2,7 +2,6 @@ export { default as Box } from './Box'
 export { default as Button } from './Button'
 export { default as Detail } from './Detail'
 export { default as Grid } from './Grid'
-export { default as Header } from './Header'
 export { default as Input } from './Input'
 export { default as Layout } from './Layout'
 export { default as Spacer } from './Spacer'
