+++ src/pages/404.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Layout } from '../components'
 
 const FourOhFour = ({ location }) => (
-  <Layout pathname={location.pathname}>
+  <Layout pageName='Not found' pathname={location.pathname}>
     <h2>404</h2>
     <p>
       There's a page here, just probably not the one you were looking
