+++ src/utils/index.js
@@ -1,2 +1,3 @@
 export { default as media } from './media'
 export { default as readingTime } from './readingTime'
+export { default as theme } from './theme'
